-added conditions for hiding components and showing depends on user is logedin or not

+++ src/components/CardFood/CardFood.js
@@ -8,17 +8,56 @@ import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import Proptype from "prop-types";
 
+function mapStateToProps(state) {
+  return {
+    user: state.user.user,
+    token: state.user.token
+  };
+}
+
+function mapActionToProps(dispatch) {
+  return bindActionCreators(
+    {
+      getItemById
+    },
+    dispatch
+  );
+}
+
 const CardFood = props => {
   let Sale = null;
   if (props.data.discountPrice > 0) {
-    Sale = <div className="menu-card__sale d-flex justify-content-center ">{props.data.discountPrice}%</div>;
+    Sale = (
+      <div className="menu-card__sale d-flex justify-content-center ">
+        {props.data.discountPrice}%
+      </div>
+    );
   }
+
+  const isLogedIn = () => {
+    let obj = {
+      deleteFood: null,
+      editFood: null,
+      favFood: null
+    };
+    if (props.user._id && props.token) {
+      obj.favFood = <i className="fa fa-heart-o" />;
+      if (props.user.roles.some(v => v === "delete"))
+        obj.deleteFood = (
+          <i onClick={() => props.delete()} className="fa fa-trash" />
+        );
+      if (props.user.roles.some(v => v === "edit"))
+        obj.editFood = <i className="fa fa-pencil" />;
+    }
+    return obj;
+  };
+
   return (
     <div className="menu-card__item">
       <div className="menu-card__edit-delete-fav">
-        <i className="fa fa-pencil" />
-        <i onClick={() => props.delete()} className="fa fa-trash" />
-        <i className="fa fa-heart-o" />
+        {isLogedIn().deleteFood}
+        {isLogedIn().editFood}
+        {isLogedIn().favFood}
       </div>
       <div className="menu-card__info">
         <div className="menu-card__image-card">
@@ -114,15 +153,7 @@ CardFood.defaultProps = {
   amount: 0
 };
 
-function mapActionToProps(dispatch) {
-  return bindActionCreators(
-    {
-      getItemById
-    },
-    dispatch
-  );
-}
 export default connect(
-  null,
+  mapStateToProps,
   mapActionToProps
 )(CardFood);
